+++ renderer_type_test.go
@@ -1,0 +1,14 @@
+package dapper_test
+
+import (
+	"testing"
+)
+
+func TestRenderer_Type(t *testing.T) {
+	test(
+		t,
+		"any",
+		nil,
+		"any(nil)",
+	)
+}